pyproject.toml
[project]
name = "loukvoicetranslate"
version = "0.1.0"
description = "Translate your voice and answer to your voice"
authors = [
    {name = "A.L",email = "abolfazlloukzade7@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "torch (>=2.8.0,<3.0.0)",
    "openai (>=1.99.9,<2.0.0)",
    "speechrecognition (>=3.14.3,<4.0.0)",
    "pydub (>=0.25.1,<0.26.0)",
    "gtts (>=2.5.4,<3.0.0)",
    "django-environ (>=0.12.0,<0.13.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
